+++ app/src/main/java/com/ian/app/muviepedia/core/data/dataSource/remote/source/tv/TvRemoteDataSource.kt
@@ -18,12 +18,4 @@ interface TvRemoteDataSource {
     suspend fun getAiringTodayTv(): DataSource<BaseResponse<TvDataResponse>>
 
     suspend fun getOnAirTv(): DataSource<BaseResponse<TvDataResponse>>
-
-    suspend fun getAiringTodayPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>>
-
-    suspend fun getOnAirPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>>
-
-    suspend fun getPopularTvPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>>
-
-    suspend fun getTopRatedPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>>
 }
