+++ app/src/main/java/com/ian/app/muviepedia/core/data/dataSource/remote/source/tv/TvRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.ian.app.muviepedia.core.data.dataSource.remote.source.tv
 
 import com.ian.app.muviepedia.BuildConfig.MOVIE_API_KEY
-import com.ian.app.muviepedia.core.data.dataSource.remote.api.ApiInterface
+import com.ian.app.muviepedia.core.data.dataSource.remote.api.TelevisionApiInterface
 import com.ian.app.muviepedia.core.data.dataSource.remote.helper.RemoteHelper
 import com.ian.app.muviepedia.core.data.dataSource.remote.model.BaseResponse
 import com.ian.app.muviepedia.core.data.dataSource.remote.model.response.DetailTvResponse
@@ -9,10 +9,11 @@ import com.ian.app.muviepedia.core.data.dataSource.remote.model.response.TvDataR
 import com.ian.app.muviepedia.core.data.model.DataSource
 import com.ian.app.muviepedia.core.data.model.RemoteBaseResult
 import com.ian.app.muviepedia.core.data.model.RemoteResult
+import com.ian.app.muviepedia.di.qualifier.TelevisionApiInterfaceQualifier
 import javax.inject.Inject
 
 class TvRemoteDataSourceImpl @Inject constructor(
-    private val api: ApiInterface,
+    @TelevisionApiInterfaceQualifier private val api: TelevisionApiInterface,
     injectedRemoteHelper: RemoteHelper
 ) : RemoteHelper by injectedRemoteHelper, TvRemoteDataSource {
     override suspend fun getDetailTv(tvID: Int): DataSource<DetailTvResponse> {
@@ -104,74 +105,6 @@ class TvRemoteDataSourceImpl @Inject constructor(
         return when (
             val data =
                 remoteWithBaseCall(api.getOnAirTvAsync(apiKey = MOVIE_API_KEY))
-        ) {
-            is RemoteBaseResult.Error -> DataSource.Error(data.exception.message ?: "")
-            is RemoteBaseResult.Success -> {
-                val result = data.data.body()
-                if (result != null) {
-                    DataSource.Success(result)
-                } else {
-                    DataSource.Error("null body")
-                }
-            }
-        }
-    }
-
-    override suspend fun getAiringTodayPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>> {
-        return when (
-            val data =
-                remoteWithBaseCall(api.pagingGetAiringTodayTvAsync(page = pageTv, apiKey = MOVIE_API_KEY))
-        ) {
-            is RemoteBaseResult.Error -> DataSource.Error(data.exception.message ?: "")
-            is RemoteBaseResult.Success -> {
-                val result = data.data.body()
-                if (result != null) {
-                    DataSource.Success(result)
-                } else {
-                    DataSource.Error("null body")
-                }
-            }
-        }
-    }
-
-    override suspend fun getOnAirPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>> {
-        return when (
-            val data =
-                remoteWithBaseCall(api.pagingGetOnAirTvAsync(page = pageTv, apiKey = MOVIE_API_KEY))
-        ) {
-            is RemoteBaseResult.Error -> DataSource.Error(data.exception.message ?: "")
-            is RemoteBaseResult.Success -> {
-                val result = data.data.body()
-                if (result != null) {
-                    DataSource.Success(result)
-                } else {
-                    DataSource.Error("null body")
-                }
-            }
-        }
-    }
-
-    override suspend fun getPopularTvPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>> {
-        return when (
-            val data =
-                remoteWithBaseCall(api.pagingGetPopularTvAsync(page = pageTv, apiKey = MOVIE_API_KEY))
-        ) {
-            is RemoteBaseResult.Error -> DataSource.Error(data.exception.message ?: "")
-            is RemoteBaseResult.Success -> {
-                val result = data.data.body()
-                if (result != null) {
-                    DataSource.Success(result)
-                } else {
-                    DataSource.Error("null body")
-                }
-            }
-        }
-    }
-
-    override suspend fun getTopRatedPaging(pageTv: Int): DataSource<BaseResponse<TvDataResponse>> {
-        return when (
-            val data =
-                remoteWithBaseCall(api.pagingGetTopRatedTvAsync(page = pageTv, apiKey = MOVIE_API_KEY))
         ) {
             is RemoteBaseResult.Error -> DataSource.Error(data.exception.message ?: "")
             is RemoteBaseResult.Success -> {
