+++ app/src/main/java/com/ian/app/muviepedia/di/module/NetworkModule.kt
@@ -1,6 +1,13 @@
 package com.ian.app.muviepedia.di.module
 
-import com.ian.app.muviepedia.core.data.dataSource.remote.api.ApiInterface
+import com.ian.app.muviepedia.core.data.dataSource.remote.api.MovieApiInterface
+import com.ian.app.muviepedia.core.data.dataSource.remote.api.MoviePaginationApiInterface
+import com.ian.app.muviepedia.core.data.dataSource.remote.api.TelevisionApiInterface
+import com.ian.app.muviepedia.core.data.dataSource.remote.api.TelevisionPaginationApiInterface
+import com.ian.app.muviepedia.di.qualifier.MovieApiInterfaceQualifier
+import com.ian.app.muviepedia.di.qualifier.MoviePagingApiInterfaceQualifier
+import com.ian.app.muviepedia.di.qualifier.TelevisionApiInterfaceQualifier
+import com.ian.app.muviepedia.di.qualifier.TelevisionPagingApiInterfaceQualifier
 import com.ian.app.muviepedia.di.scope.ApplicationScoped
 import com.squareup.moshi.Moshi
 import dagger.Module
@@ -15,7 +22,7 @@ object NetworkModule {
     private const val BASE_URL = "https://api.themoviedb.org/3/"
 
     @Provides
-    fun provideHttpClientForHomeScreen(): OkHttpClient {
+    fun provideHttpClient(): OkHttpClient {
         val logInterceptor = HttpLoggingInterceptor().apply {
             level = HttpLoggingInterceptor.Level.BODY
         }
@@ -27,15 +34,61 @@ object NetworkModule {
 
     @Provides
     @ApplicationScoped
-    fun provideApiInterfaceForHomeScreen(
+    @TelevisionApiInterfaceQualifier
+    fun provideTelevisionApiInterface(
         okHttpClient: OkHttpClient,
         moshi: Moshi
-    ): ApiInterface {
+    ): TelevisionApiInterface {
         return Retrofit.Builder()
             .client(okHttpClient)
             .addConverterFactory(MoshiConverterFactory.create(moshi))
             .baseUrl(BASE_URL)
             .build()
-            .create(ApiInterface::class.java)
+            .create(TelevisionApiInterface::class.java)
+    }
+
+    @Provides
+    @ApplicationScoped
+    @TelevisionPagingApiInterfaceQualifier
+    fun provideTelevisionPagingApiInterface(
+        okHttpClient: OkHttpClient,
+        moshi: Moshi
+    ): TelevisionPaginationApiInterface {
+        return Retrofit.Builder()
+            .client(okHttpClient)
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .baseUrl(BASE_URL)
+            .build()
+            .create(TelevisionPaginationApiInterface::class.java)
+    }
+
+    @Provides
+    @ApplicationScoped
+    @MovieApiInterfaceQualifier
+    fun provideMovieApiInterface(
+        okHttpClient: OkHttpClient,
+        moshi: Moshi
+    ): MovieApiInterface {
+        return Retrofit.Builder()
+            .client(okHttpClient)
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .baseUrl(BASE_URL)
+            .build()
+            .create(MovieApiInterface::class.java)
+    }
+
+    @Provides
+    @ApplicationScoped
+    @MoviePagingApiInterfaceQualifier
+    fun provideMoviePagingApiInterface(
+        okHttpClient: OkHttpClient,
+        moshi: Moshi
+    ): MoviePaginationApiInterface {
+        return Retrofit.Builder()
+            .client(okHttpClient)
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .baseUrl(BASE_URL)
+            .build()
+            .create(MoviePaginationApiInterface::class.java)
     }
 }
